Module twilight_model::application::command

Used for building commands to send to Discord.

Modules

permissions

Structs

BaseCommandOptionData

Data supplied to a CommandOption of type Boolean, User, Channel, Role, or Mentionable.

ChoiceCommandOptionData

Data supplied to a CommandOption of type String or Integer.

Command

Data sent to discord to create a command.

OptionsCommandOptionData

Data supplied to a CommandOption of type SubCommand or SubCommandGroup.

Enums

CommandOption

Option for a Command.

CommandOptionChoice

Specifies an option that a user must choose from in a dropdown.

CommandOptionType

Type of a CommandOption.

CommandType