Enum twilight_model::application::interaction::application_command::CommandOptionValue

pub enum CommandOptionValue {
    Attachment(Id<AttachmentMarker>),
    Boolean(bool),
    Channel(Id<ChannelMarker>),
    Focused(String, CommandOptionType),
    Integer(i64),
    Mentionable(Id<GenericMarker>),
    Number(f64),
    Role(Id<RoleMarker>),
    String(String),
    SubCommand(Vec<CommandDataOption>),
    SubCommandGroup(Vec<CommandDataOption>),
    User(Id<UserMarker>),
}

Combined value and value type for a CommandDataOption.

Variants

Attachment(Id<AttachmentMarker>)

Attachment option.

Boolean(bool)

Boolean option.

Channel(Id<ChannelMarker>)

Channel option.

Focused(String, CommandOptionType)

Focused option.

Since Discord does not validate focused fields, they are sent as strings. This means that you will not necessarily get a valid number from number options.

See Discord Docs/Autocomplete.

The actual CommandOptionType is available through the second tuple value.

Integer(i64)

Integer option.

Mentionable(Id<GenericMarker>)

Mentionable option.

Number(f64)

Number option.

Role(Id<RoleMarker>)

Role option.

String(String)

String option.

SubCommand(Vec<CommandDataOption>)

Subcommand option.

SubCommandGroup(Vec<CommandDataOption>)

Subcommand group option.

User(Id<UserMarker>)

User option.

Implementations

impl CommandOptionValue

pub const fn kind(&self) -> CommandOptionType