Struct twilight_model::application::interaction::Interaction

pub struct Interaction {

    pub app_permissions: Option<Permissions>,
    pub application_id: Id<ApplicationMarker>,
    pub channel: Option<Channel>,
    pub channel_id: Option<Id<ChannelMarker>>,
    pub data: Option<InteractionData>,
    pub entitlements: Vec<Entitlement>,
    pub guild_id: Option<Id<GuildMarker>>,
    pub guild_locale: Option<String>,
    pub id: Id<InteractionMarker>,
    pub kind: InteractionType,
    pub locale: Option<String>,
    pub member: Option<PartialMember>,
    pub message: Option<Message>,
    pub token: String,
    pub user: Option<User>,

}

Payload received when a user executes an interaction.

See Discord Docs/Interaction Object.

Fields
app_permissions: Option<Permissions>

App’s permissions in the channel the interaction was sent from.

Present when the interaction is invoked in a guild.

application_id: Id<ApplicationMarker>

ID of the associated application.

channel: Option<Channel>

The channel the interaction was invoked in.

Present on all interactions types, except Ping.

channel_id: Option<Id<ChannelMarker>>
👎Deprecated: channel_id is deprecated in the discord API and will no be sent in the future, users should use the channel field instead.

ID of the channel the interaction was invoked in.

Present on all interactions types, except Ping.

data: Option<InteractionData>

Data from the interaction.

This field present on ApplicationCommand, MessageComponent, ApplicationCommandAutocomplete and ModalSubmit interactions. The inner enum variant matches the interaction type.

entitlements: Vec<Entitlement>

For monetized apps, any entitlements for the invoking user, representing access to premium SKUs

guild_id: Option<Id<GuildMarker>>

ID of the guild the interaction was invoked in.

guild_locale: Option<String>

Guild’s preferred locale.

Present when the interaction is invoked in a guild.

id: Id<InteractionMarker>

ID of the interaction.

kind: InteractionType

Type of interaction.

locale: Option<String>

Selected language of the user who invoked the interaction.

Present on all interactions types, except Ping.

member: Option<PartialMember>

Member that invoked the interaction.

Present when the interaction is invoked in a guild.

message: Option<Message>

Message attached to the interaction.

Present on MessageComponent interactions.

token: String

Token for responding to the interaction.

user: Option<User>

User that invoked the interaction.

Present when the interaction is invoked in a direct message.

Implementations

impl Interaction

pub const fn author_id(&self) -> Option<Id<UserMarker>>

ID of the user that invoked the interaction.

This will first check for the member’s user’s ID and then, if not present, check the user’s ID.

pub const fn author(&self) -> Option<&User>

The user that invoked the interaction.

This will first check for the member’s user and then, if not present, check the user.

pub const fn is_dm(&self) -> bool

Whether the interaction was invoked in a DM.

pub const fn is_guild(&self) -> bool

Whether the interaction was invoked in a guild.