Module twilight_model::channel::message

Textual user communication method.

Re-exports

pub use self::component::Component;

pub use self::embed::Embed;

pub use self::sticker::MessageSticker;

pub use self::sticker::Sticker;

Modules

component

Interactive message elements for use with Interactions.

embed

Rich message content.

sticker

Message stickers.

Structs

AllowedMentions

Allowed mentions (pings).

Mention

Mention of a user in a message.

Message

Text message sent in a Channel.

MessageActivity

Activity associated with a message.

MessageApplication

Partial Application sent with Rich Presence-related chat embeds.

MessageCall

Information about the call in a private channel.

MessageFlags

Flags to signal state and modify the look of a message.

MessageInteraction

Associated interaction metadata.

MessageReference

Message reference struct.

MessageSnapshot

The snap-shot of a message.

Reaction

Reaction below a message.

ReactionCountDetails

Breakdown of normal and super reaction counts for the associated emoji.

RoleSubscriptionData

Information about a role subscription that created a Message.

Enums

EmojiReactionType

Type of emoji in a Reaction.

MentionType

Allowed mention type.

MessageActivityType

Activity of this message.

MessageReferenceType

The type of reference for a message.

MessageType

Type of a Message.

ReactionType

The kind of reaction.