Struct twilight_model::channel::message::allowed_mentions::AllowedMentionsBuilder

pub struct AllowedMentionsBuilder(_);

Create an AllowedMentions with a builder.

Implementations

impl AllowedMentionsBuilder

pub const fn new() -> Self

Create a new AllowedMentionsBuilder.

pub fn everyone(self) -> Self

Allow parsing of @everyone.

pub const fn replied_user(self) -> Self

When replying, whether to mention the target user.

pub fn roles(self) -> Self

Allow parsing of all roles.

pub fn role_ids(self, role_ids: impl IntoIterator<Item = Id<RoleMarker>>) -> Self

Allow parsing of specific Id<RoleMarker>s.

roles and this method are mutually exclusive. The builder will favor specific role ids.

pub fn users(self) -> Self

Allow parsing of all users.

pub fn user_ids(self, user_ids: impl IntoIterator<Item = Id<UserMarker>>) -> Self

Allow parsing of specific Id<UserMarker>s.

users and this method are mutually exclusive. The builder will favor specific user ids.

pub fn build(self) -> AllowedMentions

Build the AllowedMentions struct.

Note: This clears possible validation errors by removing data. For example, if both users and user_ids are called, the effect of users will be nullified and only the users specified with user_ids will be built.