Enum twilight_model::channel::message::MessageType

#[non_exhaustive]
pub enum MessageType {

    Regular,
    RecipientAdd,
    RecipientRemove,
    Call,
    ChannelNameChange,
    ChannelIconChange,
    ChannelMessagePinned,
    UserJoin,
    GuildBoost,
    GuildBoostTier1,
    GuildBoostTier2,
    GuildBoostTier3,
    ChannelFollowAdd,
    GuildDiscoveryDisqualified,
    GuildDiscoveryRequalified,
    GuildDiscoveryGracePeriodInitialWarning,
    GuildDiscoveryGracePeriodFinalWarning,
    ThreadCreated,
    Reply,
    ChatInputCommand,
    ThreadStarterMessage,
    GuildInviteReminder,
    ContextMenuCommand,
    AutoModerationAction,
    RoleSubscriptionPurchase,
    InteractionPremiumUpsell,
    GuildApplicationPremiumSubscription,
    Unknown(u8),

}

Type of a Message.

Refer to Discord Docs/Message Types for more information.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Regular

Regular message.

RecipientAdd

System message denoting a recipient has been added to a group.

RecipientRemove

System message denoting a recipient has been removed from a group.

Call

System message denoting a call state, e.g. missed, started.

ChannelNameChange

System message denoting a channel’s name has been changed.

ChannelIconChange

System message denoting a channel’s icon has been changed.

ChannelMessagePinned

System message denoting a message has been pinned.

UserJoin

System message denoting a member has joined a guild.

GuildBoost

System message denoting a user nitro boosted a guild.

GuildBoostTier1

System message denoting a user nitro boosted a guild to level 1.

GuildBoostTier2

System message denoting a user nitro boosted a guild to level 2.

GuildBoostTier3

System message denoting a user nitro boosted a guild to level 3.

ChannelFollowAdd

System message denoting a channel has been followed.

GuildDiscoveryDisqualified

System message denoting a guild has been disqualified for Server Discovery.

GuildDiscoveryRequalified

System message denoting a guild has been redisqualified for Server Discovery.

GuildDiscoveryGracePeriodInitialWarning

System message denoting an initial warning for Server Discovery disqualification.

GuildDiscoveryGracePeriodFinalWarning

System message denoting a final warning for Server Discovery disqualification.

ThreadCreated

Reply

Message is an inline reply.

ChatInputCommand

Message is a chat input command.

ThreadStarterMessage

GuildInviteReminder

ContextMenuCommand

AutoModerationAction

Message is an auto moderation action.

RoleSubscriptionPurchase

System message denoting a user subscribed to a role.

InteractionPremiumUpsell

System message denoting a interaction premium upsell.

GuildApplicationPremiumSubscription

System message denoting a guild application premium subscription.

Unknown(u8)

Variant value is unknown to the library.

Implementations

impl MessageType

pub const fn deletable(self) -> bool

Whether the message can be deleted, not taking permissions into account. Some message types can’t be deleted, even by server administrators.

Some message types can only be deleted with certain permissions. For example, AutoModerationAction can only be deleted if the user has the Manage Messages permission.

To check whether a message can be deleted while taking permissions into account, use deletable_with_permissions.

pub const fn deletable_with_permissions(self, permissions: Permissions) -> bool

Whether the message can be deleted, taking permissions into account. Some message types can’t be deleted, even by server administrators.

Some message types can only be deleted with certain permissions. For example, AutoModerationAction can only be deleted if the user has the Manage Messages permission.

To check whether a message can be deleted without taking permissions into account, use deletable.