Module twilight_model::gateway

Modules

connection_info

event

payload

Payloads for sending commands to and receiving events from the Discord Gateway API.

presence

Structs

CloseCodeConversionError

CloseFrame

Information about a close message.

GatewayReaction

Intents

Gateway intents.

SessionStartLimit

Current gateway session utilization status.

ShardId

Shard identifier to calculate if it receivies a given event.

ShardIdParseError

Enums

CloseCode

Gateway close event codes.

OpCode

Gateway event’s payload type.

ShardIdParseErrorType

Type of ShardIdParseError that occurred.