Enum twilight_model::gateway::event::DispatchEvent

pub enum DispatchEvent {

    AutoModerationActionExecution(AutoModerationActionExecution),
    AutoModerationRuleCreate(Box<AutoModerationRuleCreate>),
    AutoModerationRuleDelete(Box<AutoModerationRuleDelete>),
    AutoModerationRuleUpdate(Box<AutoModerationRuleUpdate>),
    BanAdd(BanAdd),
    BanRemove(BanRemove),
    ChannelCreate(Box<ChannelCreate>),
    ChannelDelete(Box<ChannelDelete>),
    ChannelPinsUpdate(ChannelPinsUpdate),
    ChannelUpdate(Box<ChannelUpdate>),
    CommandPermissionsUpdate(CommandPermissionsUpdate),
    GiftCodeUpdate,
    GuildCreate(Box<GuildCreate>),
    GuildDelete(GuildDelete),
    GuildEmojisUpdate(GuildEmojisUpdate),
    GuildIntegrationsUpdate(GuildIntegrationsUpdate),
    GuildScheduledEventCreate(Box<GuildScheduledEventCreate>),
    GuildScheduledEventDelete(Box<GuildScheduledEventDelete>),
    GuildScheduledEventUpdate(Box<GuildScheduledEventUpdate>),
    GuildScheduledEventUserAdd(GuildScheduledEventUserAdd),
    GuildScheduledEventUserRemove(GuildScheduledEventUserRemove),
    GuildStickersUpdate(GuildStickersUpdate),
    GuildUpdate(Box<GuildUpdate>),
    IntegrationCreate(Box<IntegrationCreate>),
    IntegrationDelete(IntegrationDelete),
    IntegrationUpdate(Box<IntegrationUpdate>),
    InteractionCreate(Box<InteractionCreate>),
    InviteCreate(Box<InviteCreate>),
    InviteDelete(InviteDelete),
    MemberAdd(Box<MemberAdd>),
    MemberRemove(MemberRemove),
    MemberUpdate(Box<MemberUpdate>),
    MemberChunk(MemberChunk),
    MessageCreate(Box<MessageCreate>),
    MessageDelete(MessageDelete),
    MessageDeleteBulk(MessageDeleteBulk),
    MessageUpdate(Box<MessageUpdate>),
    PresenceUpdate(Box<PresenceUpdate>),
    PresencesReplace,
    ReactionAdd(Box<ReactionAdd>),
    ReactionRemove(Box<ReactionRemove>),
    ReactionRemoveAll(ReactionRemoveAll),
    ReactionRemoveEmoji(ReactionRemoveEmoji),
    Ready(Box<Ready>),
    Resumed,
    RoleCreate(RoleCreate),
    RoleDelete(RoleDelete),
    RoleUpdate(RoleUpdate),
    StageInstanceCreate(StageInstanceCreate),
    StageInstanceDelete(StageInstanceDelete),
    StageInstanceUpdate(StageInstanceUpdate),
    ThreadCreate(Box<ThreadCreate>),
    ThreadDelete(ThreadDelete),
    ThreadListSync(ThreadListSync),
    ThreadMemberUpdate(Box<ThreadMemberUpdate>),
    ThreadMembersUpdate(ThreadMembersUpdate),
    ThreadUpdate(Box<ThreadUpdate>),
    TypingStart(Box<TypingStart>),
    UnavailableGuild(UnavailableGuild),
    UserUpdate(UserUpdate),
    VoiceServerUpdate(VoiceServerUpdate),
    VoiceStateUpdate(Box<VoiceStateUpdate>),
    WebhooksUpdate(WebhooksUpdate),

}

A dispatch event, containing information about a created guild, a member added, etc.

You can deserialize into a DispatchEvent via DispatchEventWithTypeDeserializer.

Variants

AutoModerationActionExecution(AutoModerationActionExecution)

AutoModerationRuleCreate(Box<AutoModerationRuleCreate>)

AutoModerationRuleDelete(Box<AutoModerationRuleDelete>)

AutoModerationRuleUpdate(Box<AutoModerationRuleUpdate>)

BanAdd(BanAdd)

BanRemove(BanRemove)

ChannelCreate(Box<ChannelCreate>)

ChannelDelete(Box<ChannelDelete>)

ChannelPinsUpdate(ChannelPinsUpdate)

ChannelUpdate(Box<ChannelUpdate>)

CommandPermissionsUpdate(CommandPermissionsUpdate)

GiftCodeUpdate

GuildCreate(Box<GuildCreate>)

GuildDelete(GuildDelete)

GuildEmojisUpdate(GuildEmojisUpdate)

GuildIntegrationsUpdate(GuildIntegrationsUpdate)

GuildScheduledEventCreate(Box<GuildScheduledEventCreate>)

GuildScheduledEventDelete(Box<GuildScheduledEventDelete>)

GuildScheduledEventUpdate(Box<GuildScheduledEventUpdate>)

GuildScheduledEventUserAdd(GuildScheduledEventUserAdd)

GuildScheduledEventUserRemove(GuildScheduledEventUserRemove)

GuildStickersUpdate(GuildStickersUpdate)

GuildUpdate(Box<GuildUpdate>)

IntegrationCreate(Box<IntegrationCreate>)

IntegrationDelete(IntegrationDelete)

IntegrationUpdate(Box<IntegrationUpdate>)

InteractionCreate(Box<InteractionCreate>)

InviteCreate(Box<InviteCreate>)

InviteDelete(InviteDelete)

MemberAdd(Box<MemberAdd>)

MemberRemove(MemberRemove)

MemberUpdate(Box<MemberUpdate>)

MemberChunk(MemberChunk)

MessageCreate(Box<MessageCreate>)

MessageDelete(MessageDelete)

MessageDeleteBulk(MessageDeleteBulk)

MessageUpdate(Box<MessageUpdate>)

PresenceUpdate(Box<PresenceUpdate>)

PresencesReplace

ReactionAdd(Box<ReactionAdd>)

ReactionRemove(Box<ReactionRemove>)

ReactionRemoveAll(ReactionRemoveAll)

ReactionRemoveEmoji(ReactionRemoveEmoji)

Ready(Box<Ready>)

Resumed

RoleCreate(RoleCreate)

RoleDelete(RoleDelete)

RoleUpdate(RoleUpdate)

StageInstanceCreate(StageInstanceCreate)

StageInstanceDelete(StageInstanceDelete)

StageInstanceUpdate(StageInstanceUpdate)

ThreadCreate(Box<ThreadCreate>)

ThreadDelete(ThreadDelete)

ThreadListSync(ThreadListSync)

ThreadMemberUpdate(Box<ThreadMemberUpdate>)

ThreadMembersUpdate(ThreadMembersUpdate)

ThreadUpdate(Box<ThreadUpdate>)

TypingStart(Box<TypingStart>)

UnavailableGuild(UnavailableGuild)

UserUpdate(UserUpdate)

VoiceServerUpdate(VoiceServerUpdate)

VoiceStateUpdate(Box<VoiceStateUpdate>)

WebhooksUpdate(WebhooksUpdate)

Implementations

impl DispatchEvent

pub const fn kind(&self) -> EventType

Returns the type of event that this event is.