Module twilight_model::gateway::event::shard

Structs

Connected

Indicator that a shard is now fully connected.

Connecting

Indicator that a shard is now connecting.

Disconnected

Indicator that a shard is now disconnected and may soon be reconnecting if not explicitly shutdown.

Identifying

Indicator that a shard is now identifying with the gateway to create a new session.

Payload

A payload of bytes came in through the gateway.

Reconnecting

Indicator that a shard is now reconnecting.

Resuming

Indicator that a shard is now resuming a session after a disconnect.

Enums

ShardEvent

“Meta” events about a shard’s status, not from the gateway.