Module twilight_model::gateway::payload::outgoing

Payloads that are outgoing to the Discord Gateway API.

These are sent by shards to the Discord API. Some payloads are required for the operation of a shard; these are Heartbeat, Identify, and Resume. Others - such as RequestGuildMembers - are for users to use in their own operations.

Refer to Discord Docs / Gateway Commands for Discord’s documentation about incoming and outgoing events.

Re-exports

pub use self::identify::Identify;

pub use self::request_guild_members::RequestGuildMembers;

pub use self::resume::Resume;

pub use self::update_presence::UpdatePresence;

pub use self::update_voice_state::UpdateVoiceState;

Modules

identify

request_guild_members

resume

update_presence

update_voice_state

Structs

Heartbeat