Struct twilight_model::gateway::Intents

pub struct Intents { /* private fields */ }

Gateway intents.

Developers must specify intents when connecting to the gateway. The intents specified correspond with the events received. To specify multiple intents, create a union using the | operator. See Discord Docs/Gateway Intents.

Implementations

impl Intents

pub const GUILDS: Self = _

Guilds intent.

Event(s) received:

GUILD_CREATE

GUILD_UPDATE

GUILD_DELETE

GUILD_ROLE_CREATE

GUILD_ROLE_UPDATE

GUILD_ROLE_DELETE

CHANNEL_CREATE

CHANNEL_UPDATE

CHANNEL_DELETE

CHANNEL_PINS_UPDATE

THREAD_CREATE

THREAD_UPDATE

THREAD_DELETE

THREAD_LIST_SYNC

THREAD_MEMBER_UPDATE

THREAD_MEMBERS_UPDATE

STAGE_INSTANCE_CREATE

STAGE_INSTANCE_UPDATE

STAGE_INSTANCE_DELETE

pub const GUILD_MEMBERS: Self = _

Guild members intent.

This intent is privileged. See Discord Docs/Privileged Intents.

Event(s) received:

GUILD_MEMBER_ADD

GUILD_MEMBER_UPDATE

GUILD_MEMBER_REMOVE

THREAD_MEMBERS_UPDATE

pub const GUILD_BANS: Self = _

👎Deprecated since 0.14.3: use the GUILD_MODERATION intent instead

Guild bans intent.

Event(s) received:

GUILD_BAN_ADD

GUILD_BAN_REMOVE

pub const GUILD_MODERATION: Self = _

Guild moderation intent.

Event(s) received:

GUILD_AUDIT_LOG_ENTRY_CREATE

GUILD_BAN_ADD

GUILD_BAN_REMOVE

pub const GUILD_EMOJIS_AND_STICKERS: Self = _

Guild emojis and stickers intent.

Event(s) received:

GUILD_EMOJIS_UPDATE

GUILD_STICKERS_UPDATE

pub const GUILD_INTEGRATIONS: Self = _

Guild integrations intent.

Event(s) received:

GUILD_INTEGRATIONS_UPDATE

INTEGRATION_CREATE

INTEGRATION_UPDATE

INTEGRATION_DELETE

pub const GUILD_WEBHOOKS: Self = _

Guild webhooks intent.

Event(s) received:

WEBHOOKS_UPDATE

pub const GUILD_INVITES: Self = _

Guild invites intent.

Event(s) received:

INVITE_CREATE

INVITE_DELETE

pub const GUILD_VOICE_STATES: Self = _

Guild voice states intent.

Event(s) received:

VOICE_STATE_UPDATE

pub const GUILD_PRESENCES: Self = _

Guild presences intent.

This intent is privileged. See Discord Docs/Privileged Intents.

Event(s) received:

PRESENCE_UPDATE

pub const GUILD_MESSAGES: Self = _

Guild messages intent.

Event(s) received:

MESSAGE_CREATE

MESSAGE_UPDATE

MESSAGE_DELETE

MESSAGE_DELETE_BULK

pub const GUILD_MESSAGE_REACTIONS: Self = _

Guild message reactions intent.

Event(s) received:

MESSAGE_REACTION_ADD

MESSAGE_REACTION_REMOVE

MESSAGE_REACTION_REMOVE_ALL

MESSAGE_REACTION_REMOVE_EMOJI

pub const GUILD_MESSAGE_TYPING: Self = _

Guild message typing intent.

Event(s) received:

TYPING_START

pub const DIRECT_MESSAGES: Self = _

Direct messages intent.

Event(s) received:

MESSAGE_CREATE

MESSAGE_UPDATE

MESSAGE_DELETE

MESSAGE_DELETE_BULK

This is different from the GUILD_MESSAGES intent in that the bot will receive message events from locations other than guilds.

pub const DIRECT_MESSAGE_REACTIONS: Self = _

Direct message reactions intent.

Event(s) received:

MESSAGE_REACTION_ADD

MESSAGE_REACTION_REMOVE

MESSAGE_REACTION_REMOVE_ALL

MESSAGE_REACTION_REMOVE_EMOJI

This is different from the GUILD_MESSAGE_REACTIONS event in that the bot will receive message reaction events from locations other than guilds.

pub const DIRECT_MESSAGE_TYPING: Self = _

Direct message typing intent.

Event(s) received:

TYPING_START

This is different from the GUILD_MESSAGE_TYPING intent in that the bot will receive typing start events from locations other than guilds.

pub const MESSAGE_CONTENT: Self = _

Message content intent.

This intent is privileged. See Discord Docs/Privileged Intents.

This intent allows you to receive the contents of all messages.

pub const GUILD_SCHEDULED_EVENTS: Self = _

Guild scheduled events intent.

Event(s) received:

GUILD_SCHEDULED_EVENT_CREATE

GUILD_SCHEDULED_EVENT_UPDATE

GUILD_SCHEDULED_EVENT_DELETE

GUILD_SCHEDULED_EVENT_USER_ADD

GUILD_SCHEDULED_EVENT_USER_REMOVE

pub const AUTO_MODERATION_CONFIGURATION: Self = _

Auto moderation configuration intent.

Event(s) received:

AUTO_MODERATION_RULE_CREATE

AUTO_MODERATION_RULE_DELETE

AUTO_MODERATION_RULE_UPDATE

pub const AUTO_MODERATION_EXECUTION: Self = _

Auto moderation execution event.

Event(s) received:

AUTO_MODERATION_ACTION_EXECUTION

pub const fn empty() -> Self

Returns an empty set of flags.

pub const fn all() -> Self

Returns the set containing all flags.

pub const fn bits(&self) -> u64

Returns the raw value of the flags currently stored.

pub const fn from_bits(bits: u64) -> Option<Self>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u64) -> Self

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const unsafe fn from_bits_unchecked(bits: u64) -> Self

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

Safety

The caller of the bitflags! macro can chose to allow or disallow extra bits for their bitflags type.

The caller of from_bits_unchecked() has to ensure that all bits correspond to a defined flag or that extra bits are valid for this bitflags type.

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: Self) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: Self) -> bool

Returns true if all of the flags in other are contained within self.

pub fn insert(&mut self, other: Self)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: Self)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: Self)

Toggles the specified flags in-place.

pub fn set(&mut self, other: Self, value: bool)

Inserts or removes the specified flags depending on the passed value.

pub const fn intersection(self, other: Self) -> Self

Returns the intersection between the flags in self and other.

Specifically, the returned set contains only the flags which are present in both self and other.

This is equivalent to using the & operator (e.g. ops::BitAnd), as in flags & other.

pub const fn union(self, other: Self) -> Self

Returns the union of between the flags in self and other.

Specifically, the returned set contains all flags which are present in either self or other, including any which are present in both (see Self::symmetric_difference if that is undesirable).

This is equivalent to using the | operator (e.g. ops::BitOr), as in flags | other.

pub const fn difference(self, other: Self) -> Self

Returns the difference between the flags in self and other.

Specifically, the returned set contains all flags present in self, except for the ones present in other.

It is also conceptually equivalent to the “bit-clear” operation: flags & !other (and this syntax is also supported).

This is equivalent to using the - operator (e.g. ops::Sub), as in flags - other.

pub const fn symmetric_difference(self, other: Self) -> Self

Returns the symmetric difference between the flags in self and other.

Specifically, the returned set contains the flags present which are present in self or other, but that are not present in both. Equivalently, it contains the flags present in exactly one of the sets self and other.

This is equivalent to using the ^ operator (e.g. ops::BitXor), as in flags ^ other.

pub const fn complement(self) -> Self

Returns the complement of this set of flags.

Specifically, the returned set contains all the flags which are not set in self, but which are allowed for this type.

Alternatively, it can be thought of as the set difference between Self::all() and self (e.g. Self::all() - self)

This is equivalent to using the ! operator (e.g. ops::Not), as in !flags.