Module twilight_model::guild

Group of Channels and Users with additional moderation options.

Guilds allow, for example, assigning Roles to Members to limit their Permissions globally, or per Channel.

Re-exports

pub use self::widget::GuildWidget;

Modules

audit_log

Audit Logs, created whenever an administrative action is performed within a guild.

auto_moderation

Types for Auto Moderation.

invite

onboarding

Types for guild onboarding.

scheduled_event

Types for interacting with scheduled events.

template

widget

Structs

AfkTimeout

Duration of a user being AFK before being timed out from a voice channel.

Ban

Emoji

Guild

GuildInfo

GuildIntegration

GuildPreview

GuildPrune

IntegrationAccount

IntegrationApplication

Member

User that is in a Guild.

MemberFlags

Flags associated with a Member.

PartialGuild

PartialMember

Permissions

Role

RoleFlags

The flags for a role represented as a bitfield.

RolePosition

Data used to update the positions of roles.

RoleTags

Tags that a Role has.

SystemChannelFlags

UnavailableGuild

VanityUrl

Information about a guild’s vanity URL setting.

Enums

DefaultMessageNotificationLevel

ExplicitContentFilter

GuildFeature

Special and optional guild features.

GuildIntegrationType

Special and optional guild features.

IntegrationExpireBehavior

Behavior to perform when the user’s integration expires.

MfaLevel

NSFWLevel

PremiumTier

VerificationLevel