Module twilight_model::guild::audit_log

Audit Logs, created whenever an administrative action is performed within a guild.

For additional information refer to Discord Docs/Audit Logs.

Structs

AffectedRole

Minimal amount of information about an affected role.

AuditLog

Paginated audit log entries with additional information.

AuditLogEntry

Entry in an AuditLog possibly containing a number of detailed changes.

AuditLogGuildIntegration

Information about a guild integration provided in an audit log.

AuditLogOptionalEntryInfo

Additional information for certain AuditLogEventTypes.

Enums

AuditLogChange

Individual change within an AuditLogEntry.

AuditLogChangeKey

Type of AuditLogChange.

AuditLogChangeTypeValue

Value of a change which may be one of multiple types.

AuditLogEventType

Action to cause an AuditLogEntry.