Module twilight_model::guild::auto_moderation

Types for Auto Moderation.

Auto Moderation is a feature which allows each guild to set up rules that trigger based on some criteria. For example, a rule can trigger whenever a message contains a specific keyword.

Rules can be configured to automatically execute actions whenever they trigger. For example, if a user tries to send a message which contains a certain keyword, a rule can trigger and block the message before it is sent.

Structs

AutoModerationAction

An action which will execute whenever a rule is triggered.

AutoModerationActionMetadata

Additional metadata needed during execution for a specific AutoModerationActionType.

AutoModerationRule

Configured auto moderation rule.

AutoModerationTriggerMetadata

Additional data used to determine whether a rule should be triggered.

Enums

AutoModerationActionType

Type of AutoModerationAction.

AutoModerationEventType

Indicates in what event context a rule should be checked.

AutoModerationKeywordPresetType

Internally pre-defined wordsets which will be searched for in content.

AutoModerationTriggerType

Characterizes the type of content which can trigger the rule.