Module twilight_model::oauth::scope

Known list of available OAuth2 scopes.

Refer to Discord Docs/OAuth 2 Scopes for a complete up-to-date list.

Constants

ACTIVITIES_READ

Allows your app to fetch data from a user’s “Now Playing/Recently Played” list.

ACTIVITIES_WRITE

Allows your app to update a user’s activity

APPLICATIONS_BUILDS_READ

Allows your app to read build data for a user’s applications.

APPLICATIONS_BUILDS_UPLOAD

Allows your app to upload/update builds for a user’s applications.

APPLICATIONS_COMMANDS

Allows your app to use commands in a guild.

APPLICATIONS_COMMANDS_PERMISSIONS_UPDATE

Allows your app to update permissions for its commands in a guild a user has permissions to.

APPLICATIONS_COMMANDS_UPDATE

Allows your app to update its commands using a Bearer token.

APPLICATIONS_ENTITLEMENTS

Allows your app to read entitlements for a user’s applications

APPLICATIONS_STORE_UPDATE

Allows your app to read and update store data (SKUs, store listings, achievements, etc.) for a user’s applications.

BOT

For oauth2 bots, this puts the bot in the user’s selected guild by default

CONNECTIONS

Allows /users/@me/connections to return linked third-party accounts

DM_CHANNELS_READ

Allows your app to see information about the user’s DMs and group DMs

EMAIL

Enables GET /users/@me returning an email.

GDM_JOIN

Allows your app to join users to a group DM.

GUILDS

Allows GET /users/@me/guilds to return basic information about all of a user’s guilds.

GUILDS_JOIN

Allows GET /guilds/{guild.id}/members/{user.id} to be used for joining users to a guild.

GUILDS_MEMBERS_READ

Allows GET /users/@me/guilds/{guild.id}/member to return a user’s member information in a guild.

IDENTIFY

Allows GET /users/@me, but without the user’s email.

MESSAGES_READ

For local RPC server API access, this allows you to read messages from all client channels (otherwise restricted to channels/guilds your app creates).

RELATIONSHIPS_READ

Allows your app to know a user’s friends and implicit relationships.

ROLE_CONNECTIONS_WRITE

Allows your app to update a user’s connection and metadata for the app.

RPC

For local RPC server access, this allows you to control a user’s local Discord client.

RPC_ACTIVITIES_WRITE

For local rpc server access, this allows you to update a user’s activity

RPC_NOTIFICATIONS_READ

For local RPC server access, this allows you to receive notifications pushed out to the user.

RPC_VOICE_READ

For local RPC server access, this allows you to read a user’s voice settings and listen for voice events.

RPC_VOICE_WRITE

For local RPC server access, this allows you to update a user’s voice settings.

VOICE

Allows your app to connect to voice on the user’s behalf and see all the voice members.

WEBHOOK_INCOMING

This generates a webhook that is returned in the oauth token response for authorization code grants.