Module twilight_model::channel

Re-exports

pub use self::message::Message;

pub use self::stage_instance::StageInstance;

pub use self::webhook::Webhook;

Modules

forum

message

Textual user communication method.

permission_overwrite

stage_instance

thread

webhook

Structs

Attachment

Channel

Channel to send messages in, call with other users, organize groups, and more.

ChannelFlags

ChannelMention

FollowedChannel

An object indicating that following a news channel was successful.

Enums

ChannelType

VideoQualityMode

WebhookType