Module twilight_model::channel::message

Textual user communication method.

Re-exports

pub use self::component::Component;

pub use self::embed::Embed;

pub use self::sticker::Sticker;

Modules

component

Interactive message elements for use with Interactions.

embed

Rich message content.

sticker

Message stickers.

Structs

AllowedMentions

Allowed mentions (pings).

Mention

Mention of a user in a message.

Message

Text message sent in a Channel.

MessageActivity

Activity associated with a message.

MessageApplication

Partial Application sent with Rich Presence-related chat embeds.

MessageFlags

Flags to signal state and modify the look of a message.

MessageInteraction

Associated interaction metadata.

MessageReference

Message reference struct.

Reaction

Reaction below a message.

RoleSubscriptionData

Information about a role subscription that created a Message.

Enums

MentionType

Allowed mention type.

MessageActivityType

Activity of this message.

MessageType

Type of a Message.

ReactionType

Type of Reaction.