Module twilight_model::gateway::payload

Payloads for sending commands to and receiving events from the Discord Gateway API.

An example of an incoming event might be RoleCreate while an outgoing event might be UpdateVoiceState.

Refer to Discord Docs / Commands and Events for Discord’s documentation about incoming and outgoing events.

Modules

incoming

Payloads that are incoming from the Discord Gateway API.

outgoing

Payloads that are outgoing to the Discord Gateway API.