Module twilight_model::gateway::payload::incoming

Payloads that are incoming from the Discord Gateway API.

These are sent by the Discord Gateway API to shards. The only payload required for the operation of a shard is Ready; all other payloads are for users to consume in their own operations based on the Intents they have set

Refer to Discord Docs / Gateway Events for Discord’s documentation about incoming and outgoing events.

Re-exports

pub use self::invite_create::InviteCreate;

pub use self::reaction_remove_emoji::ReactionRemoveEmoji;

Modules

invite_create

Gateway event payload when an invite is created.

reaction_remove_emoji

Structs

AutoModerationActionExecution

Message has been blocked by AutoMod according to a rule.

AutoModerationRuleCreate

Auto moderation rule has been created.

AutoModerationRuleDelete

Auto moderation rule has been deleted.

AutoModerationRuleUpdate

Auto moderation rule has been updated.

BanAdd

BanRemove

ChannelCreate

ChannelDelete

ChannelPinsUpdate

ChannelUpdate

CommandPermissionsUpdate

GuildAuditLogEntryCreate

The inner value of the GuildAuditLogEntryCreate variant of the Event enum.

GuildCreate

GuildDelete

GuildEmojisUpdate

GuildIntegrationsUpdate

GuildScheduledEventCreate

Sent when a guild scheduled event is created. The inner payload is a GuildScheduledEvent.

GuildScheduledEventDelete

Sent when a guild scheduled event is deleted. The inner payload is a GuildScheduledEvent.

GuildScheduledEventUpdate

Sent when a guild scheduled event is updated. The inner payload is a GuildScheduledEvent.

GuildScheduledEventUserAdd

Sent when a user has subscribed to a guild scheduled event.

GuildScheduledEventUserRemove

Sent when a user has unsubscribed from a guild scheduled event.

GuildStickersUpdate

GuildUpdate

Hello

IntegrationCreate

IntegrationDelete

IntegrationUpdate

InteractionCreate

InviteDelete

MemberAdd

MemberChunk

MemberRemove

MemberUpdate

MessageCreate

MessageDelete

MessageDeleteBulk

MessageUpdate

PresenceUpdate

User’s presence was updated.

ReactionAdd

ReactionRemove

ReactionRemoveAll

Ready

RoleCreate

RoleDelete

RoleUpdate

StageInstanceCreate

StageInstanceDelete

StageInstanceUpdate

ThreadCreate

ThreadDelete

ThreadListSync

ThreadMemberUpdate

ThreadMembersUpdate

ThreadUpdate

TypingStart

UnavailableGuild

UserUpdate

VoiceServerUpdate

VoiceStateUpdate

WebhooksUpdate