Module twilight_model::gateway::payload::incoming

Payloads that are incoming from the Discord Gateway API.

These are sent by the Discord Gateway API to shards. The only payload required for the operation of a shard is Ready; all other payloads are for users to consume in their own operations based on the Intents they have set

Refer to Discord Docs / Gateway Events for Discord’s documentation about incoming and outgoing events.

Re-exports

pub use self::invite_create::InviteCreate;

pub use self::reaction_remove_emoji::ReactionRemoveEmoji;

Modules

invite_create

Gateway event payload when an invite is created.

reaction_remove_emoji

Structs

BanAdd

BanRemove

ChannelCreate

ChannelDelete

ChannelPinsUpdate

ChannelUpdate

GuildCreate

GuildDelete

GuildEmojisUpdate

GuildIntegrationsUpdate

GuildStickersUpdate

GuildUpdate

IntegrationCreate

IntegrationDelete

IntegrationUpdate

InteractionCreate

InviteDelete

MemberAdd

MemberChunk

MemberRemove

MemberUpdate

MessageCreate

MessageDelete

MessageDeleteBulk

MessageUpdate

PresenceUpdate

ReactionAdd

ReactionRemove

ReactionRemoveAll

Ready

RoleCreate

RoleDelete

RoleUpdate

StageInstanceCreate

StageInstanceDelete

StageInstanceUpdate

ThreadCreate

ThreadDelete

ThreadListSync

ThreadMemberUpdate

ThreadMembersUpdate

ThreadUpdate

TypingStart

UnavailableGuild

UserUpdate

VoiceServerUpdate

VoiceStateUpdate

WebhooksUpdate