Module twilight_model::guild

Group of Channels and Users with additional moderation options.

Guilds allow, for example, assigning Roles to Members to limit their Permissions globally, or per Channel.

Advanced user’s may look inside the member module for custom Member deserializers.

Modules

audit_log

Audit Logs, created whenever an administrative action is performed within a guild.

auto_moderation

Types for Auto Moderation.

invite

member

Mostly internal custom serde deserializers.

scheduled_event

Types for interacting with scheduled events.

template

Structs

Ban

Emoji

Guild

GuildInfo

GuildIntegration

GuildPreview

GuildPrune

GuildWidget

IntegrationAccount

IntegrationApplication

Member

User that is in a Guild.

PartialGuild

PartialMember

Permissions

Role

RoleTags

Tags that a Role has.

SystemChannelFlags

UnavailableGuild

VanityUrl

Information about a guild’s vanity URL setting.

Enums

DefaultMessageNotificationLevel

ExplicitContentFilter

GuildFeature

Special and optional guild features.

IntegrationExpireBehavior

Behavior to perform when the user’s integration expires.

MfaLevel

NSFWLevel

PremiumTier

VerificationLevel