Enum twilight_model::guild::audit_log::AuditLogChangeKey

#[non_exhaustive]
pub enum AuditLogChangeKey {

    AfkChannelId,
    AfkTimeout,
    Allow,
    ApplicationId,
    Archived,
    Asset,
    AutoArchiveDuration,
    Available,
    AvatarHash,
    BannerHash,
    Bitrate,
    ChannelId,
    Code,
    Color,
    CommandId,
    CommunicationDisabledUntil,
    Deaf,
    DefaultAutoArchiveDuration,
    DefaultMessageNotifications,
    Deny,
    Description,
    DiscoverySplashHash,
    EnableEmoticons,
    EntityType,
    ExpireBehavior,
    ExpireGracePeriod,
    ExplicitContentFilter,
    FormatType,
    GuildId,
    Hoist,
    IconHash,
    Id,
    ImageHash,
    Invitable,
    InviterId,
    Location,
    Locked,
    MaxAge,
    MaxUses,
    Mentionable,
    MfaLevel,
    Mute,
    Name,
    Nick,
    Nsfw,
    NsfwLevel,
    OwnerId,
    PermissionOverwrites,
    Permissions,
    Position,
    PreferredLocale,
    PrivacyLevel,
    PruneDeleteDays,
    PublicUpdatesChannelId,
    RateLimitPerUser,
    Region,
    RoleAdded,
    RoleRemoved,
    RulesChannelId,
    SplashHash,
    Status,
    SystemChannelId,
    Tags,
    Temporary,
    Topic,
    Type,
    UnicodeEmoji,
    UserLimit,
    Uses,
    VanityUrlCode,
    VerificationLevel,
    WidgetChannelId,
    WidgetEnabled,

}

Type of AuditLogChange.

For additional information refer to Discord Docs/Audit Log Change Key.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AfkChannelId

AFK voice channel for a guild.

AfkTimeout

Timeout to cause a user to be moved to an AFK voice channel.

Allow

Allowed permissions of a permission overwrite target.

ApplicationId

ID of an application.

Archived

Thread was archived or unarchived.

Asset

Asset of a sticker.

Empty string.

AutoArchiveDuration

Auto archive duration of a thread.

Available

Availability of a sticker.

AvatarHash

Hash of an avatar.

BannerHash

Hash of a guild banner.

Bitrate

Bitrate of an audio channel.

ChannelId

Channel for an invite code.

Code

Code of an invite.

Color

Color of a role.

CommandId

Permissions for a command were updated.

CommunicationDisabledUntil

Member timeout state changed.

Deaf

Whether a user is guild deafened.

DefaultAutoArchiveDuration

Default auto archive duration for new threads.

DefaultMessageNotifications

Default message notification level for a guild.

Deny

Denied permissions of a permission overwrite target.

Description

Description of a guild.

DiscoverySplashHash

Hash of a guild’s discovery splash.

EnableEmoticons

Whether emoticons are enabled.

EntityType

Entity type of guild scheduled event was changed.

ExpireBehavior

Behavior of the expiration of an integration.

ExpireGracePeriod

Grace period of the expiration of an integration.

ExplicitContentFilter

Explicit content filter level of a guild.

FormatType

Format type of a sticker.

GuildId

Guild that a sticker is in.

Hoist

Whether a role is hoisted.

IconHash

Hash of a guild icon.

Id

ID of an entity.

ImageHash

Hash of a guild scheduled event cover.

Invitable

Invitable state of a private thread.

InviterId

ID of the user who created an invite.

Location

Channel ID for a scheduled event changed.

Locked

Thread was locked or unlocked.

MaxAge

Maximum age of an invite.

MaxUses

Maximum uses of an invite.

Mentionable

Whether a role can be mentioned in a message.

MfaLevel

Multi-Factor Authentication level required of a guild’s moderators.

Mute

Whether a user is guild muted.

Name

Name of an entity such as a channel or role.

Nick

Nickname of a member.

Nsfw

Whether a channel is NSFW.

NsfwLevel

NSFW level of a guild.

OwnerId

ID of the owner of a guild.

PermissionOverwrites

Permission overwrites on a channel changed.

Permissions

Default permissions of a role.

Position

Position of an entity such as a channel or role.

PreferredLocale

Preferred locale of a guild.

PrivacyLevel

Privacy level of a stage instance.

PruneDeleteDays

Number of days’ worth of inactivity for a guild prune.

PublicUpdatesChannelId

ID of a guild’s public updates channel.

RateLimitPerUser

Ratelimit per user in a textual channel.

Region

Region of a guild changed.

RoleAdded

Role added to a user.

RoleRemoved

Role removed from a user.

RulesChannelId

ID of a guild’s rules channel.

SplashHash

Hash of a guild’s splash.

Status

Status of guild scheduled event was changed.

SystemChannelId

ID of a guild’s system channel.

Tags

Related emoji of a sticker.

Temporary

Whether an invite is temporary.

Topic

Topic of a textual channel.

Type

Type of a created entity.

UnicodeEmoji

Role unicode emoji.

UserLimit

Maximum number of users in a voice channel.

Uses

Number of uses of an invite.

VanityUrlCode

Code of a guild’s vanity invite.

VerificationLevel

Required verification level of new members in a guild.

WidgetChannelId

Channel ID of a widget.

WidgetEnabled

Whether a widget is enabled.

Implementations

impl AuditLogChangeKey

pub const fn name(self) -> &'static str

Raw name of the key.

The raw names of keys are in snake_case form.

Examples

Check the names of the Allow and BannerHash keys:

use twilight_model::guild::audit_log::AuditLogChangeKey;

assert_eq!("allow", AuditLogChangeKey::Allow.name());
assert_eq!("banner_hash", AuditLogChangeKey::BannerHash.name());