Enum twilight_model::guild::audit_log::AuditLogChangeKey

#[non_exhaustive]
pub enum AuditLogChangeKey {

    AfkChannelId,
    AfkTimeout,
    Allow,
    ApplicationId,
    AvatarHash,
    BannerHash,
    Bitrate,
    ChannelId,
    Code,
    Color,
    Deaf,
    DefaultMessageNotifications,
    Deny,
    Description,
    DiscoverySplashHash,
    EnableEmoticons,
    ExpireBehavior,
    ExpireGracePeriod,
    ExplicitContentFilter,
    Hoist,
    IconHash,
    Id,
    InviterId,
    MaxAge,
    MaxUses,
    Mentionable,
    MfaLevel,
    Mute,
    Name,
    Nick,
    NsfwLevel,
    OwnerId,
    Permissions,
    Position,
    PreferredLocale,
    PrivacyLevel,
    PruneDeleteDays,
    PublicUpdatesChannelId,
    RateLimitPerUser,
    RoleAdded,
    RoleRemoved,
    RulesChannelId,
    SplashHash,
    SystemChannelId,
    Temporary,
    Topic,
    Type,
    UserLimit,
    Uses,
    VanityUrlCode,
    VerificationLevel,
    WidgetChannelId,
    WidgetEnabled,

}

Type of AuditLogChange.

For additional information refer to Discord Docs/Audit Log Change Key.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AfkChannelId

AFK voice channel for a guild.

AfkTimeout

Timeout to cause a user to be moved to an AFK voice channel.

Allow

Allowed permissions of a permission overwrite target.

ApplicationId

ID of an application.

AvatarHash

Hash of an avatar.

BannerHash

Hash of a guild banner.

Bitrate

Bitrate of an audio channel.

ChannelId

Channel for an invite code.

Code

Code of an invite.

Color

Color of a role.

Deaf

Whether a user is guild deafened.

DefaultMessageNotifications

Default message notification level for a guild.

Deny

Denied permissions of a permission overwrite target.

Description

Description of a guild.

DiscoverySplashHash

Hash of a guild’s discovery splash.

EnableEmoticons

Whether emoticons are enabled.

ExpireBehavior

Behavior of the expiration of an integration.

ExpireGracePeriod

Grace period of the expiration of an integration.

ExplicitContentFilter

Explicit content filter level of a guild.

Hoist

Whether a role is hoisted.

IconHash

Hash of a guild icon.

Id

ID of an entity.

InviterId

ID of the user who created an invite.

MaxAge

Maximum age of an invite.

MaxUses

Maximum uses of an invite.

Mentionable

Whether a role can be mentioned in a message.

MfaLevel

Multi-Factor Authentication level required of a guild’s moderators.

Mute

Whether a user is guild muted.

Name

Name of an entity such as a channel or role.

Nick

Nickname of a member.

NsfwLevel

NSFW level of a guild.

OwnerId

ID of the owner of a guild.

Permissions

Default permissions of a role.

Position

Position of an entity such as a channel or role.

PreferredLocale

Preferred locale of a guild.

PrivacyLevel

Privacy level of a stage instance.

PruneDeleteDays

Number of days’ worth of inactivity for a guild prune.

PublicUpdatesChannelId

ID of a guild’s public updates channel.

RateLimitPerUser

Ratelimit per user in a textual channel.

RoleAdded

Role added to a user.

RoleRemoved

Role removed from a user.

RulesChannelId

ID of a guild’s rules channel.

SplashHash

Hash of a guild’s splash.

SystemChannelId

ID of a guild’s system channel.

Temporary

Whether an invite is temporary.

Topic

Topic of a textual channel.

Type

Type of a created entity.

UserLimit

Maximum number of users in a voice channel.

Uses

Number of uses of an invite.

VanityUrlCode

Code of a guild’s vanity invite.

VerificationLevel

Required verification level of new members in a guild.

WidgetChannelId

Channel ID of a widget.

WidgetEnabled

Whether a widget is enabled.

Implementations

impl AuditLogChangeKey

pub const fn name(self) -> &'static str

Raw name of the key.

The raw names of keys are in snake_case form.

Examples

Check the names of the Allow and BannerHash keys:

use twilight_model::guild::audit_log::AuditLogChangeKey;

assert_eq!("allow", AuditLogChangeKey::Allow.name());
assert_eq!("banner_hash", AuditLogChangeKey::BannerHash.name());