Module twilight_model::id

Type-safe IDs for each resource to avoid mixing the IDs of resources like channels and guilds.

serde

These IDs support deserializing from both integers and strings and serialize into strings.

Structs

ApplicationId

AttachmentId

AuditLogEntryId

ChannelId

EmojiId

GenericId

GuildId

IntegrationId

MessageId

RoleId

UserId

WebhookId