Module twilight_model::id::marker

Markers for various resource types, such as channels or users.

Markers themselves perform no logical action, and are only used to ensure that IDs of incorrect types aren’t used. If IDs were only 64-bit integers then a role’s ID may be erroneously used in the place of where a user’s ID is required; by using markers it can be ensured that only an ID with a RoleMarker can be used where a role’s ID is required.

Structs

ApplicationMarker

Marker for application IDs.

AttachmentMarker

Marker for attachment IDs.

AuditLogEntryMarker

Marker for audit log entry IDs.

AutoModerationRuleMarker

Marker for auto moderation rule IDs.

ChannelMarker

Marker for channel IDs.

CommandMarker

Marker for command IDs.

CommandVersionMarker

Marker for command versions.

EmojiMarker

Marker for emoji IDs.

GenericMarker

Marker for generic IDs.

GuildMarker

Marker for guild IDs.

IntegrationMarker

Marker for integration IDs.

InteractionMarker

Marker for interaction IDs.

MessageMarker

Marker for message IDs.

OauthSkuMarker

Marker for OAuth SKU IDs.

OauthTeamMarker

Marker for OAuth team IDs.

RoleMarker

Marker for role IDs.

RoleSubscriptionSkuMarker

Marker for SKU IDs.

ScheduledEventEntityMarker

Marker for scheduled event entity IDs.

ScheduledEventMarker

Marker for scheduled event IDs.

StageMarker

Marker for stage IDs.

StickerBannerAssetMarker

Marker for sticker banner asset IDs.

StickerMarker

Marker for sticker IDs.

StickerPackMarker

Marker for sticker pack IDs.

StickerPackSkuMarker

Marker for sticker pack SKU IDs.

TagMarker

Marker for forum tag IDs.

UserMarker

Marker for user IDs.

WebhookMarker

Marker for webhook IDs.